+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.3.0 - Unreleased
 - added `global`and `nonlocal`statements
 - added single-line if/elif/else, for, while and def (e.g. `def f(x): return 42` without newline)
+- enabled line continuation: Function calls, expressions or other code can now span multiple lines as long as there is an unclosed parenthesis `()` or `[]` or `{}`
 
 ## 1.2.1
 - improve package description and fix linter warnings